Se agrega MVC para completar booking

+++ Controllers/ServiceController.php
@@ -21,16 +21,14 @@
 
         public function ShowListView()
         {
- 
-            $serviceList = $this->serviceDAO->getAll();
-
             require_once(VIEWS_PATH."example.php");
         }
 
-        public function Add($startDate, $endDate, $status){
-            $user = $_SESSION["loggedUser"];
+        public function Add($startDate, $endDate, $status, $keeper = NULL){
+            $userController = new UserController();
+            $user = $keeper != null ? $userController->userDAO->GetById($keeper) : $_SESSION["loggedUser"];
             $serviceList = $this->serviceDAO->getAll();
-            if($endDate < $startDate){
+            if($keeper == null && $endDate < $startDate){
                 $message = 'You cannot set the end date to before the start date';
                 echo "<script>alert('$message');</script>";
             }else{
@@ -47,30 +45,42 @@
                     $service->setEndDate($endDate);
                     $service->setStatus($status);
                     $this->serviceDAO->Add($service);
-
-                    $message = 'Your availability has been successfully set';
-                    echo "<script>alert('$message');</script>";
+                    
+                    if($keeper == null){
+                        $message = 'Your availability has been successfully set';
+                        echo "<script>alert('$message');</script>";
+                    }
                 }else if($flag == 1){
                     $message = 'Some of the dates entered had already been loaded. Please enter different dates.'; 
                     echo "<script>alert('$message');</script>";
                 } 
-            }   
-            $serviceList = $this->serviceDAO->getAll();
-            $this->ShowAvailabilityView();
+            } 
+            if($keeper != NULL){
+                $keeperController = new KeeperController();
+                $keeperController->ShowListView();
+            }else{
+                $serviceList = $this->serviceDAO->getAll();
+                $this->ShowAvailabilityView();
+            }
+            
         }
 
-        public function ModifyService($service, $startDate, $endDate){
+        public function ModifyService($service, $startDate, $endDate, $keeper){
             $serviceStartDate = $service->getStartDate();
             $serviceEndDate = $service->getEndDate();
             $this->Remove($service->getId());
-            if($serviceStartDate == $startDate && $serviceEndDate == $endDate){
-                $this->Add($startDate, $endDate, 'pending');
-            }
+            $this->Add($startDate, $endDate, 'pending', $keeper);
             if($serviceStartDate < $startDate){
-                $this->Add($serviceStartDate, date('Y-m-d', strtotime($startDate . '-1 day')), 'available');
+                $date=date_create($startDate);
+                date_add($date,date_interval_create_from_date_string("-1 days"));
+                $date = date_format($date, "Y-m-d");
+                $this->Add($serviceStartDate, $date, 'available', $keeper);
             }
             if($serviceEndDate > $endDate){
-                $this->Add(strtotime($endDate . '+1 day'), $serviceEndDate, 'available');
+                $date=date_create($endDate);
+                date_add($date,date_interval_create_from_date_string("+1 days"));
+                $date = date_format($date, "Y-m-d");
+                $this->Add($date, $serviceEndDate, 'available', $keeper);
             }
         }
 
@@ -78,9 +88,8 @@
         {
             require_once(VIEWS_PATH."validate-session.php");
             
-            $this->servicerDAO->Remove($id);
+            $this->serviceDAO->Remove($id);
 
-            $this->ShowAvailabilityView();
         }
 
         public function getServices(){
